requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
openai==1.54.3
requests==2.31.0
openai-agents==0.1.0

